Move widget controls into own directory
