+++ src/app/[handle]/components/widget/widget-controls/widget-delete-button.tsx
